changed a lot to the design of the site.
Added a new index page that will be the main entry point of the site. Old index.html has been renamed to lopez_math_index.html.  Added a header and the style for it is stored in header.css and the javascript for it is stored in index_script.js but will change that to header.js.  The header.css and (soon to be) header.js can be used on every page and the html can just be copy-pasted.